use parentheses instead of parenthesis when relevant
