+++ crates/jarl-core/src/analyze/parenthesized_expression.rs
@@ -1,5 +1,5 @@
 use crate::checker::Checker;
-use crate::lints::base::unnecessary_parenthesis::unnecessary_parenthesis::unnecessary_parenthesis;
+use crate::lints::base::unnecessary_parentheses::unnecessary_parentheses::unnecessary_parentheses;
 use crate::rule_set::Rule;
 use air_r_syntax::RParenthesizedExpression;
 
@@ -8,7 +8,7 @@ pub fn parenthesized_expression(
     checker: &mut Checker,
 ) -> anyhow::Result<()> {
     if checker.is_rule_enabled(Rule::UnnecessaryParenthesis) {
-        checker.report_diagnostic(unnecessary_parenthesis(r_expr)?);
+        checker.report_diagnostic(unnecessary_parentheses(r_expr)?);
     }
 
     Ok(())
