+++ docs/changelog.md
@@ -16,7 +16,7 @@
   * `notin` (#459, @Yousa-Mirage)
   * `pipe_consistency` (#482)
   * `pipe_return` (#502)
-  * `unnecessary_parenthesis`
+  * `unnecessary_parentheses`
 
 * Jarl is now available on PyPI under the name `jarl-linter`, enabling its
   installation via `uv`, `pipx`, and other tools (#466).
