+++ docs/rules.qmd
@@ -131,7 +131,7 @@ dat <- as.data.frame(
     c("unexplained_suppression", "comments", "❌", ""),
     c("unmatched_range_suppression", "comments", "❌", ""),
     c("unnecessary_nesting", "readability", "✅", "Disabled by default"),
-    c("unnecessary_parenthesis", "readability", "❌", ""),
+    c("unnecessary_parentheses", "readability", "❌", ""),
     c("unreachable_code", "readability, suspicious", "❌", ""),
     c("unused_function", "correctness", "❌", ""),
     c("unused_suppression", "comments", "❌", ""),
