+++ docs/rules/unnecessary_parenthesis.md
@@ -1,4 +1,4 @@
-# unnecessary_parenthesis
+# unnecessary_parentheses
 
 ::: {.callout-note title="Added in 0.6.0" .low-opacity}
 :::
